+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### [Report an Issue](https://github.com/atlassian/atlascode/issues)
 
+## What's new in 3.8.12
+
+### Bug Fixes
+
+- Fixed issue in create issue page when site select is empty after changing issue type
+
 ## What's new in 3.8.11
 
 ### Features